Limit ops-console-status to current node (SCRD-8346)
The ops-console-status.yml playbook was using a URL on the VIP to
determine the status, which does not reflect the state of the service on
the each node; if the ops console was not running on a particular node,
the service was still considered to be running since the VIP was routing
the request to another node (on which the service was still running).

Change the URL to query for status to be the bind address on the local
host on which the playbook is running against.

Change-Id: I875b36b0155afc40bae3abc626a9c5b3e44cdec5